LayoutTests/http/tests/security/contentSecurityPolicy/worker-importscripts-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/resources/post-message.js because it does not appear in the script-src directive of the Content Security Policy.
PASS result is "importScripts blocked: NetworkError (DOM Exception 19):  A network error occurred."
PASS successfullyParsed is true

TEST COMPLETE
